fix: 네트워크 메시지 오류 수정
