+++ EzyBook/Infrastructure/Networking/DefaultNetworkService.swift
@@ -1,16 +1,22 @@
 //
-//  DefaultHttpClient.swift
+//  DefaultNetworkService.swift
 //  EzyBook
 //
-//  Created by youngkyun park on 5/21/25.
+//  Created by youngkyun park on 5/13/25.
 //
 
 import Foundation
 import Alamofire
 
-final class DefaultHttpClient: HttpClient {
+final class DefaultNetworkService: NetworkService {
+
+    private let decodingManager: ResponseDecoder
     
-    func request<R: NetworkRouter>(_ router: R) async throws -> Data {
+    init(decodingManager: ResponseDecoder) {
+        self.decodingManager = decodingManager
+    }
+    
+    func fetchData<T: Decodable & EntityConvertible, E: StructEntity, R: NetworkRouter>(dto: T.Type ,_ router: R) async throws -> E where T.E == E {
         
         let urlRequest: URLRequest
         do {
@@ -19,17 +25,23 @@ final class DefaultHttpClient: HttpClient {
             throw APIError(localErrorType: .missingEndpoint)
         }
         
-        //Alamofire에서 Combine이나 async/await을 위해 제공하는 구조
+        
         let response = await AF.request(urlRequest)
+            .validate(statusCode: 200...299)
             .serializingData()
             .response
-        
         switch response.result {
         case .success(let data):
-            return data
+            let decodedResult = decodingManager.decode(data: data, type: dto)
+            switch decodedResult {
+            case .success(let decodedDTO):
+                return decodedDTO.toEntity()
+            case .failure(let decodeError):
+                throw decodeError
+            }
         case .failure(let afError):
             let statusCode = response.response?.statusCode
-            let responseData = response.data //에러메시지
+            let responseData = response.data
             if let code = statusCode {
                 throw APIError(statusCode: code, data: responseData)
             } else {
@@ -38,5 +50,7 @@ final class DefaultHttpClient: HttpClient {
             }
         }
     }
+    
 }
+
